requirements.txt
polygon-api-client
sqlalchemy
alembic
click
loguru
yfinance
psycopg2-binary
python-dotenv
pandas_market_calendars
